Rearrange shortcuts for toggling floating mode
I used to toggle full floating mode for all windows on workspace but I
realized that I only really wanted one window to float. I discovered
that I can toggle floating per window which is something I did not
realize even though I had this option in my config this whole time.
However it was bound to a combination that I wanted to use for something
else. In the end, I assigned the command to toggle per window floating
to what was previously used to toggle full floating layout. And the
shortcut used for toggling per window floating is now used for
displaying file browser in rofi, the initial change that inspired all of
these updates.